fix: improve podcast error handling and validation
- Add duration validation: must be positive and <= 24 hours (86400 seconds)
- Add required field validation: title and audio_url
- Input sanitization with .strip() on all text fields
- Replace database error messages with user-friendly messages
- Improve error logging with %s format for better debugging
- Users no longer see raw database errors, only helpful messages
- Backend logs full exception details for troubleshooting

+++ models.py
@@ -1288,7 +1288,7 @@ class Podcast(db.Model):
   guest = db.Column(db.String(255))  # Guest name
   audio_url = db.Column(db.String(500), nullable=False)
   thumbnail_url = db.Column(db.String(500))
-  duration = db.Column(db.Integer)  # Duration in seconds
+  duration = db.Column(db.BigInteger)  # Duration in seconds
   episode_number = db.Column(db.Integer)
   season_number = db.Column(db.Integer)
   category = db.Column(db.String(100))
